:truck: renamed main to _core
