+++ bionty/gene/__init__.py
@@ -1,2 +1,2 @@
-from .main import Gene
+from ._core import Gene
 from ._query import Biomart, Mygene
